requirements.txt
langchain
langchain-openai
langchain-community
chromadb
sentence-transformers
pypdf
python-dotenv
langchain-huggingface
langchain-chroma
